pyproject.toml
[project]
name = "stillwater"
version = "0.1.0"
description = "Stillwater - A meditation and mindfulness app"
readme = "README.md"
authors = [
    { name = "Daniel Siegle", email = "daniel.siegle@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "PyJWT[crypto]>=2.9.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.12",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[project.scripts]
stillwater = "stillwater:main"

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
# B008: FastAPI's Depends() in function signatures is idiomatic and intentional.
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
